server/ml/requirements.txt
sentence-transformers==2.7.0
faiss-cpu==1.8.0
pandas==2.2.2
numpy==1.26.4
tqdm==4.66.4
pymongo>=4.6
python-dotenv>=1.0
